requirements.txt
alembic==1.13.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.62.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.0.1
numpy==1.26.4
packaging==23.2
pandas==2.1.4
pillow==10.4.0
plotly==6.2.0
pluggy==1.6.0
propelauth-fastapi==4.2.8
propelauth-py==4.2.8
protobuf==4.25.8
psycopg2-binary==2.9.9
pyarrow==21.0.0
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.28.1
tenacity==8.5.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
websockets==15.0.1
zipp==3.23.0
